Crate nest

Nest is a crate to start protyping and making 2d graphics with as little effort as possible. It's powerful combinators assist in making ordered transformations on different levels to build larger and larger shapes.

This library does perform well, but not as well as rendering manually with custom shaders. For almost any 2d application this performance will be sufficient, but this library is not designed for high-performance drawing. If you are drawing more than a million polygons per frame, you may not want to use this library, but profile as desired. We are still committed to improving the performance given our API as much as possible.

Example

extern crate nest;
use nest::*;
 
fn main() {
    let mut app = Window::new("Example", 640, 480).expect("error: failed to open window");
 
    while !app.poll_events().any(|e| e == Event::Closed) {
        app.draw(rect([-0.5, -0.5], [0.5, 0.5]).translate((-0.1, -0.1)).combine(
            rect([-0.8, -0.8], [0.3, 0.3])).translate([0.1, 0.1]).rotate(0.5));
    }
}

Modules

error

Structs

Color
Color in (red, green, blue, alpha) tuple form.

Combine
Combine represents the combination of two shapes.

Error
The Error type.

Frame
Frame can be used to make multiple separate draw calls in one frame. When it drops/exits scope, it automatically updates the window with the frame, but it can also be finished by calling the method finish().

Positions
Three positions which form a matrix for shader purposes

Rect
Two points make a rectangle.

RendTri
Renderable triangle which includes color and texture information.

Rotate
Rotate represents a shape which has been rotated.

RotateIter
Iterator which is produced by Rotate

Translate
Translate represents a shape which has been translated.

TranslateIter
Iterator which is produced by Translate

Tri
A triangle primitive which enters the shader pipeline as a single vertex and is the only primitive in nest

Window
Represets a window with OpenGL context.

Enums

ErrorKind
The kind of an error.

Event
Represents events passed form Window.

Key

KeyState

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Shape
Trait for structs to be drawn with Frame::draw

ToSeconds
Trait for things which can be converted into seconds. This should not be used for high-precision accuracy.

Functions

rect
Takes two points and creates a rectangle from those two points.

Type Definitions

Result
Convenient wrapper around std::Result.