Struct nest::Tri

pub struct Tri {
    pub positions: Positions,
    pub texcoords: Positions,
    pub color: [f32; 4],
}

A triangle primitive which enters the shader pipeline as a single vertex and is the only primitive in nest

Fields
positions: Positions

The three space vertices of the triangles

texcoords: Positions

The three texture coordinates of the above vertices

color: [f32; 4]

The color of this triangle.

Methods

impl Tri

fn new<P: Into<Point2<f32>> + Copy, T: Into<Point2<f32>> + Copy, C: Into<Color>>(
positions: [P; 3],
texcoords: [T; 3],
color: C
) -> Tri

Create a new triangle with points and tex coordinates specified

fn new_pos<P: Into<Point2<f32>> + Copy>(positions: [P; 3]) -> Tri

Create a new triangle with points with coordinates to create a single-color triangle