Struct nest::Window

pub struct Window { /* fields omitted */ }

Represets a window with OpenGL context.

Example

use nest::*;
use std::iter::empty;
 
fn main() {
    let mut app = Window::new("Window Example", 640, 480).unwrap();
 
    while !app.poll_events().any(|e| e == Event::Closed) {
        app.draw(empty());
    }
}

Methods

impl Window

fn new<S: Into<String>>(title: S, width: u32, height: u32) -> Result<Self>

Create a new Window with a Title, and size (width / height) specified in pixels.

Pacnic

This method will panic if there are any issues with creating the window or compiling the shaders. See the glium library for more specifics on window creation.

Example

extern crate nest;
use nest::Window;

let mut app = Window::new("Hello World", 640, 480);

fn load_image<P: AsRef<Path>>(&self, path: P) -> Result<Rc<Texture2d>>

Load an image from a file to be drawn by Frame::draw_image(...).

Parameters

path - a path pointing to an image file

Example

let mut app = Window::new("Window Example", 640, 480).unwrap();
let pic = app.load_image("res/city.jpg").unwrap();

fn clear_color<C: Into<Color>>(&mut self, color: C)

Change the color which the screen is cleared with between frames.

Example

let mut app = Window::new("Window Example", 640, 480).unwrap();
app.clear_color(Color::RED);

fn poll_events(&mut self) -> IntoIter<Event>

Poll the window for events.

Example

let mut app = Window::new("Window Example", 640, 480).unwrap();

for event in app.poll_events() {
    match event {
        // Close if they close the window or hit escape.
        Event::Closed | Event::KeyboardInput(KeyState::Pressed, Some(Key::Escape)) => return,
        // Print "Space!" if they hit space.
        Event::KeyboardInput(KeyState::Pressed, Some(Key::Space)) => println!("Space!"),
        _ => {}
    }
}

fn frame<'a>(&'a self) -> Frame<'a>

Clears the screen and gets a Frame which can be used to make multiple draw calls to in one frame. When dropped, Frame automatically draws the frame.

Example

let mut app = Window::new("Window Example", 640, 480).unwrap();

while !app.poll_events().any(|e| e == Event::Closed) {
    let mut frame = app.frame();
    frame.draw(rect([0.0, 0.0], [1.0, 1.0]));
    frame.finish();
}

fn draw<S>(&mut self, shape: S) where
S: Shape,

Clears the frame, draws the shape, and updates the window.